requirements.txt
langchain==1.0.2
langchain-core>=1.0.0,<2.0.0
langchain-ollama==1.0.0
chainlit
python-dotenv
chromadb
PyPDF2
sentence-transformers
networkx
